FlaskServer/requirements.txt
grpcio-tools
grpcio
googleapis-common-protos
flask
mysql-connector-python